smartsheet-processes/requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.23.0
requests-toolbelt==0.9.1
six==1.14.0
smartsheet-python-sdk==2.86.0
urllib3==1.25.8
